requirements.txt
aiogram>=3.0.0
python-dotenv
pytz
requests
openai>=1.0.0
feedparser
beautifulsoup4
python-dateutil
aiohttp